+++ src/err/mod.rs
@@ -25,22 +25,21 @@
 
 mod macros;
 
-use once_cell::sync::Lazy;
-use std::{
+use core::{
     any::{Any, TypeId},
-    collections::HashSet,
-    env,
-    error::Error,
     fmt::{Debug, Display},
-    sync::Mutex,
 };
+
+use std::{collections::HashSet, error::Error, sync::Mutex};
+
+use once_cell::sync::Lazy;
 
 // avoid out-of-order printing
 static LOG_LK: Mutex<()> = Mutex::new(());
 
 /// `INFO` or `ERROR`, if mismatch, default to `INFO`
 pub static LOG_LEVEL: Lazy<String> = Lazy::new(|| {
-    if let Ok(l) = env::var("RUC_LOG_LEVEL") {
+    if let Ok(l) = std::env::var("RUC_LOG_LEVEL") {
         if "ERROR" == l {
             return "ERROR".to_owned();
         }
@@ -49,7 +48,7 @@ pub static LOG_LEVEL: Lazy<String> = Lazy::new(|| {
 });
 
 /// Custom Result
-pub type Result<T> = std::result::Result<T, Box<dyn RucError>>;
+pub type Result<T> = core::result::Result<T, Box<dyn RucError>>;
 
 /// the major trait defination
 pub trait RucError: Display + Debug + Send {
@@ -58,7 +57,8 @@ pub trait RucError: Display + Debug + Send {
 
     /// type ids of errors of each level(from top to bottom).
     fn type_ids(&self) -> Vec<TypeId> {
-        let mut res = vec![self.type_id()];
+        let mut res = Vec::new();
+        res.push(self.type_id());
         while let Some(c) = self.cause() {
             res.push(c.type_id());
         }
@@ -184,10 +184,6 @@ pub trait RucError: Display + Debug + Send {
             )
         }
 
-        #[cfg(target_arch = "wasm32")]
-        let pid = 0;
-
-        #[cfg(not(target_arch = "wasm32"))]
         let pid = std::process::id();
 
         // can not call `p` in the inner,
@@ -229,7 +225,7 @@ impl<T, E: Debug + Display + Send> RucResult<T, E> for Option<T> {
 }
 
 impl<T, E: Debug + Display + Send, ERR: Error> RucResult<T, E>
-    for std::result::Result<T, ERR>
+    for core::result::Result<T, ERR>
 {
     #[inline(always)]
     fn c(self, msg: SimpleMsg<E>) -> Result<T> {
@@ -261,7 +257,7 @@ impl<E: Debug + Display + Send + 'static> SimpleError<E> {
 }
 
 impl<E: Debug + Display + Send + 'static> Display for SimpleError<E> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         write!(f, "{}", self.generate_log(None))
     }
 }
@@ -342,7 +338,7 @@ impl<E: Debug + Display + Send + 'static> SimpleMsg<E> {
 
 impl<E: Debug + Display + Send + 'static> Display for SimpleMsg<E> {
     #[cfg(feature = "ansi")]
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         write!(
             f,
             "{0}{4}{5}file: {1}{4}{5}line: {2}{4}{6}column: {3}",
@@ -357,7 +353,7 @@ impl<E: Debug + Display + Send + 'static> Display for SimpleMsg<E> {
     }
 
     #[cfg(not(feature = "ansi"))]
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         write!(
             f,
             "\x1b[01m{0}\x1b[00m{4}{5}\x1b[01mfile:\x1b[00m {1}{4}{5}\x1b[01mline:\x1b[00m {2}{4}{6}\x1b[01mcolumn:\x1b[00m {3}",
@@ -383,7 +379,7 @@ impl<E: Debug + Display + Send + 'static> From<SimpleMsg<E>>
 #[inline(always)]
 #[cfg(target_os = "linux")]
 fn get_pidns(pid: u32) -> Result<String> {
-    std::fs::read_link(format!("/proc/{}/ns/pid", pid))
+    std::fs::read_link(format!("/proc/{pid}/ns/pid"))
         .c(crate::d!())
         .map(|p| {
             p.to_string_lossy()