+++ src/lib.rs
@@ -11,16 +11,16 @@ pub mod common;
 pub mod err;
 
 #[cfg(feature = "cmd")]
-#[cfg(not(target_arch = "wasm32"))]
 pub mod cmd;
 
 #[cfg(feature = "ssh")]
-#[cfg(not(target_arch = "wasm32"))]
 pub mod ssh;
 
 #[cfg(feature = "uau")]
-#[cfg(not(target_arch = "wasm32"))]
 #[cfg(target_os = "linux")]
 pub mod uau;
 
 pub use err::*;
+
+#[cfg(feature = "crypto")]
+pub mod crypto;